Time: 0 ms (100.00%) | Memory: 22 MB (16.20%) - LeetSync

+++ 98-validate-binary-search-tree/validate-binary-search-tree.cpp
@@ -12,7 +12,7 @@
 class Solution {
 public:
 
-bool isBST(TreeNode* root, long long min, long long max){
+bool isBST(TreeNode* root, long min, long max){
     //base case
     if(root == NULL)
         return true;
